+++ src/controller/post-controller.js
@@ -2,12 +2,10 @@ import {
   addPost,
   updatePost,
   deletePost,
-  addComment,
   addLikes,
   getUserLike,
   updateTypePost,
   deleteLikes,
-  getAllComments,
 } from '../model/user-post.js';
 import { currentUser } from '../model/auth-user.js';
 
@@ -86,35 +84,13 @@ export const eventDeletePost = (event) => {
   }
 };
 
-export const eventAddComment = (event) => {
-  // agregar comentario db
-  event.preventDefault();
-  const btnAddComment = event.target;
-  const postId = btnAddComment.closest('.card-post').id;
-  const comment = btnAddComment.closest('.card-post').querySelector('#text-comment');
-  const objComment = {
-    idPost: postId,
-    textComment: comment.value,
-    user: currentUser().displayName,
-  };
-  if (comment.value !== '') {
-    addComment(objComment)
-      .then((doc) => {
-        comment.value = '';
-        console.log('comentario agregado exitosamente', doc);
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  }
-};
-
 export const addAndDeleteLikes = (event) => {
   event.preventDefault();
   const btnLike = event.target;
   const idPost = btnLike.closest('.card-post').id;
   getUserLike(idPost, currentUser().uid)
     .then((likes) => {
+      console.log(likes);
       if (!likes.empty) {
         console.log('el usuario ya no puede dar like');
         likes.forEach((doclike) => {
@@ -145,24 +121,4 @@ export const eventChangeTypePost = (event) => {
         console.log(error);
       });
   }
-};
-
-export const eventGetAllComments = (event) => {
-  const btnGetAllComments = event.target;
-  const idPost = btnGetAllComments.closest('.card-post').id;
-  getAllComments()
-    .then((querySnapshot) => {
-      querySnapshot.forEach((doc) => {
-        console.log(doc.id);
-        console.log(idPost);
-        console.log(doc.data().idPostComment);
-
-        if (doc.id === idPost) {
-          console.log(doc.data());
-        }
-      });
-    })
-    .catch((error) => {
-      console.log(error);
-    });
 };
