+++ src/css/styles.css
@@ -388,8 +388,8 @@ p {
     padding-left: 4px;
   }
 
-  .like-post i{
-    color: #ff0000;
+  .i-dislike{
+    color: #808080;
   }
 
   .update-post i{
@@ -567,7 +567,7 @@ nav#superior ul li:nth-child(3) {
     padding-left: 4px;
   }
 
-  .like-post i{
+  .i-like{
     color: #ff0000;
   }
