+++ src/model/auth-user.js
@@ -23,10 +23,13 @@ export const currentUser = () => firebase.auth().currentUser;
 
 export const getUserData = idUser => firebase.firestore().collection('users').doc(idUser).get();
 
-export const updateProfile = (idProfile, newTextProfileType, newTextProfileName) => {
+export const updateProfile = (idProfile, newTextProfileType, newTextProfileName, newaboutme,
+  newcountry) => {
   const result = firebase.firestore().collection('users').doc(idProfile).update({
     displayName: newTextProfileName,
     type: newTextProfileType,
+    aboutme: newaboutme,
+    country: newcountry,
   });
   return result;
 };
